+++ lib/train.ml
@@ -211,7 +211,7 @@ let%track3_sexp round_robin_dry_run ~num_streams jitbs ~dry_sync : unit =
   loop jitbs;
   if !pos % num_streams <> 0 then dry_sync (!pos % num_streams)
 
-let set_virtual (a : Tn.t) = Tn.update_memory_mode a Virtual 29
+let set_virtual (a : Tn.t) = Tn.update_memory_mode a (Virtual { is_constant = false }) 29
 
 let every_non_literal_on_host =
   Tensor.iter_embedded ~f:(fun a ->
@@ -267,7 +267,8 @@ let%track3_sexp parallel_update (type buffer_ptr dev runner event)
     let occupancy ~name:_ ~src_n:_ = true in
     Array.mapi ctxs ~f:(fun dst_n ctx ->
         if occupancy_dst ~dst_n then
-          snd @@ Backend.(link_batch ctx @@ compile_batch ~occupancy Idx.Empty grad_merges)
+          let empty_ctx = Ir.Low_level.{ computations = Hashtbl.create (module Ir.Tnode) } in
+          snd @@ Backend.(link_batch ctx @@ compile_batch empty_ctx ~occupancy Idx.Empty grad_merges)
         else [||])
   in
   (* We can cache scheduling, because merging and copying does not depend on static indexing. *)